Refactored process for checking commands to run
